+++ app/cart/page.tsx
@@ -1,0 +1,12 @@
+import Cart from "@/components/payments/Cart"
+import { useFetchCourses } from "@/hooks/courses/useFetchCourse"
+import { getCartCourses } from "@/lib/firebase/queries/getCartCourses"
+
+export default async function CartPage() {
+    //  const courses = useFetchCourses()
+
+    const courseQuery = useFetchCourses()
+    const course = courseQuery.data
+
+    return <Cart courses={course} />
+}
