+++ components/StatsSection.tsx
@@ -1,3 +1,4 @@
+"use client"
 import React, { useState, useEffect } from "react"
 import { useInView } from "react-intersection-observer"
 import styled from "styled-components"
