+++ components/Testimonials.tsx
@@ -1,3 +1,4 @@
+"use client"
 import React, { useState } from "react"
 import Avatar from "@mui/material/Avatar"
 import Card from "@mui/material/Card"
